added ui and discounts
