+++ src/main/java/com/booleanuk/core/Customer.java
@@ -1,6 +1,7 @@
 package com.booleanuk.core;
 
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.util.List;
 
 public class Customer {
@@ -39,4 +40,5 @@ public class Customer {
     {
         return basket.getListOfItemsInBasket().get(i);
     }
+
 }
